drivers: net: Renaming ti816x->emac
