+++ src/drivers/net/ti816x/Mybuild
@@ -5,10 +5,10 @@ module ti816x {
 	option number prep_buff_cnt	= 128 /* the number of prepared buffers for rxing */
 
 	@IncludeExport(path="drivers/net")
-	source "ti816x.h"
+	source "emac.h"
 
-	source "ti816x.c"
-	source "mdio.c"
+	source "emac.c"
+	source "emac_mdio.c"
 
 	option number speed = 100
